Drop options for MongoClient.
There is no need to be that pedantic. Those options are for fine-tuning
only and can be misleading. Someone who wants to set those options would
clearly understand some advanced mechanism of MongoDB and can be
expected to know how **kwargs works in the skeleton. I don't think we
are making the life of this person harder by hiding those options. See
pep20.