Make the url of table_of_contents.xml tweakable
That url is the only link to the eurostat website that was hardcoded in
dlstats. Now that we can change that link in the configuration file, we
can serve a fake eurostat website to dlsats in order to ease the
development process. Specifically, the developer needs to be able to
create revisions to some time series on demand.