Fix codes in the categories collection
This wasn't tested but we need to quickly find a fix for that.
Currently, the code were simply absent from the category collection, not
providing references between the categories and the series.

Alternatively, we could leave the situation as it was before this commit
and provide additional queries in ulstats to deal with that. Those
queries would be crazy expensive. Following that route, we should use
map reduce and cache some of those queries. It is too early to take a
definitive decision about that. Let's go forward and not be (too much)
afraid about refactoring later.