Initial commit for World Bank data
UNlike the name stated in the author field, this commit has been mostly
done by Salimeh Monirijavid.
This commit seems to be somehow failing in the middle of the list of
series. We are investigating that issue. A fix will be issued in the
days to come. Also, we should be using bulk_inserts instead of one shots
"fire and forget" inserts. This does not seem to work as of now.

Many methods are also missing (wrt dlstats.fetchers._skeleton.Skeleton()).