Hotfix for imports in the tests
Previously, all the tests were importing system-wide modules. This is
obviously a bad practice that stems from a bug in setuptools. In other
words, setuptools fails to properly implement PEP328 and, as a
consequence, give the absolute prority to systemwide packages in
sys.path.
Long story short, python distribution process remains a mess, but
everything is manageable.

In terms of process, I am directly pushing this to Widukind/dlstats
because it does not introduce new logic, does not break any part of the
API and is very small.